requirements.txt
py-cord
pymysql
aiomysql
apscheduler
youtube_dl
bs4
pynacl
html5lib
pycryptodome
requests
itunes-app-scraper-dmi
Pillow
python-dateutil
cairosvg
